created CustomerReportModal
